♻️ Refactor MQTT
