+++ src/components/AircraftLevelPopup.tsx
@@ -3,7 +3,7 @@ import { observer } from 'mobx-react-lite';
 import React from 'react';
 import { Button } from 'react-bootstrap';
 
-import { changeFlightLevelOfAircraft } from '../mqtt';
+import { changeFlightLevelOfAircraft, handlePublishPromise } from '../mqtt/publishers';
 import { configurationStore, cwpStore } from '../state';
 import type AircraftModel from '../model/AircraftModel';
 
@@ -84,7 +84,9 @@ export default observer(function AircraftLevelPopup(properties: { aircraft: Airc
       setLocalAssignedFlightLevel(stringFlightLevel);
     } else {
       setAssignedFlightLevel(stringFlightLevel);
-      changeFlightLevelOfAircraft(controlledBy, assignedFlightId, stringFlightLevel);
+      handlePublishPromise(
+        changeFlightLevelOfAircraft(controlledBy, assignedFlightId, stringFlightLevel),
+      );
     }
     close();
   };
