+++ src/mqtt/index.ts
@@ -1,0 +1,3 @@
+import * as mqtt from 'mqtt';
+
+export default mqtt;
